# Optimise ext table dumping