When Mnesia is dumping transactions logs with operations on ext tables, it first groups operations in a commit by the ext storage engine. The current implementation folds over the list with `orddict:append/3`, which is an `O(n^2)` algorithm. This is especially problematic because there are typically only a handful of ext storage engines used by any database, while a transaction can contain an arbitrary large number of operations. So the `orddict` built this way will typically have very few keys with very long list of values stored under them.

It is faster to fold over the list in reverse order and prepend each operation to the list of already collected operations. A further optimisation is using a map instead of an orddict. And finally, since external storage engines with `disc_only_copies` semantics only need to be dumped at startup, we shall filter out operations on those right when building the map for all other dump types.

Together these optimisations yield a massive speed up when dumping large commits. In our scenario, we have 1 external storage engine with `disc_only_copies` semantics, and some transactions may contain ~1-2 million updates to these tables. Dumping these transactions takes hours. I made some measurements on my laptop to illustrate the effect (only considering the time it takes to group and iterate over the transactions, not the time to dump them to disc):

`InitBy` | `disc_copies` engines | `disc_only_copes` engines | Original | Optimised |
| --- | --- | --- | ---: | ---: |
| `user` | 1 with 30k ops | 1 with 30k ops | 6,605 ms | 12 ms |
| `user` | 0 | 1 with 30k ops | 2,182 ms | 2 ms |
| `user` | 0 | 1 with 10k ops | 159 ms | <1 ms |
| `user` | 0 | 1 with 1000 ops | 1,683 us | 47 us |
| `user` | 0 | 1 with 100 ops | 24 us | 19 us |
| `user` | 0 | 1 with 10 ops | 2 us | 1 us |
| `startup` | 1 with 30k ops | 1 with 30k ops | 5,337 ms | 18 ms |
| `startup` | 0 | 1 with 30k ops | 1,821 ms | 15 ms |
| `startup` | 0 | 1 with 10k ops | 161 ms | 2 ms |
| `startup` | 0 | 1 with 1000 ops | 2,172 us | 205 us |
| `startup` | 0 | 1 with 100 ops | 43 us | 18 us |
| `startup` | 0 | 1 with 10 ops | 2 us | 4 us |

My benchmarking methodology wasn't very advanced (doing a single run only, no consideration for gc etc.), but it looks obvious that the optimisations are useful, except for the smallest test scenarios where the two methods performed within the margin of error.