@dszoboszlay 2019-01-08
I was thinking a bit more about the problem, and now I believe the whole grouping by external engine bit is unnecessary. This would allow us to further simplify the code as seen in 48432cf.

However, that solution would change the order in which `insert_op/5` calls are made if there are more than one external engines involved in the commit (the `insert_op/5` calls from different engines would be interleaved, but each engine would receive its respective `insert_op/5` calls in the original order). I don't see any problems with this change, however I'm not confident enough with my understanding of Mnesia internals to definitely say it's correct.

So please consider carefully reviewing 48432cf as well!