@richcarl 2019-01-22
Well, Daniel sits at the table behind me, and this is a real thing that bit us when we had some unusually large transactions, so we just want this merged upstream. Paging @mikpe as well if he wants to weigh in.