@dszoboszlay 2019-01-22
I'm quite confident the change as seen in this PR is correct, because it does exactly the same thing as the original code, only in an optimised way.

But I would really like to hear your opinion on whether the alternative I proposed in 31542e4 could work? Is there any inherent reason for dispatching ops grouped by backend engine or is it only because that's how Mnesia does it with native `disc_copies` and `disc_only_copies` tables (where the ops are already grouped, so there's no performance penalty for this strategy)?