@uwiger 2019-01-22
I can't come up with a reason why the approach in [31542e4](https://github.com/erlang/otp/commit/31542e40d6ddf2be61979218e38dfa9f2829bd05) shouldn't work.

The performance gain is impressive, so the change seems like a good one.