@dszoboszlay 2019-01-22
I opened #2102 from 31542e4, if you think that patch is correct, it should be a better option than 28c4ec1 (this PR).