@mikpe 2019-01-22
Answering @richcarl 's page, my only comment is that we have been bitten badly by the O(N^2) behaviour of the current code on several occasions, so eliminating that is a top priority.